requirements.txt
aws-cdk-lib==2.127.0
constructs>=10.3.0,<11.0.0
aws-cdk.aws-glue-alpha==2.127.0a0
boto3==1.34.64
pyyaml==6.0.1
python-dotenv==1.0.1
